Handle assembly loading errors at composition time
The default implementation of MEF's DirectoryCatalog does not
handle ReflectionTypeLoadExceptions that occur when assemblies
in the specified directory are loaded.  This causes the full
assembly loading process to stop, meaning that no rules will
be found.  When running Script Analyzer in a folder with a lot
of DLLs, this type of error becomes more likely.  This change
introduces a new SafeDirectoryCatalog type which is more resilient
to this problem.

Also, there is a check that determines whether any rules have
been loaded before continuing.  This check was prone to throwing
NullReferenceExceptions when MEF composition fails.  I changed
that logic to be more resilient to potential null values.