Creates a workaround the failing tests.
We have set the CustomRulePath parameter in the cmdlet implementation in LibraryUsage.tests.ps1 to "string[]" type whereas in the c# implementation it is of "string" type. If we set the CustomRulePath parameter here to "string", then the library usage test fails when run as an administrator. We want to note that the library usage test doesn't fail when run as a non-admin user. Even if we create a "[string[]]" type object and pass it to Initialize method,  the tests fail to run as an admin with the following error message.

Assert failed on "Initialize" with "7" argument(s): "Test failed due to terminating error: The module was expected to contain an assembly manifest. (Exception from HRESULT: 0x80131018)