Fix severity filtering while invoking script analyzer
Filters the rules based on their severity before passing them the AST for
analysis.

Prior to the fix, the severity filter was applied on diagnostic records.
This was inefficient as Invoke-ScriptAnalyzer would run all the rules
irrespective of the given severity and then filter the diagnostic records
to extract the results that corresponded to the given severity.