Fix parse error caused by missing dsc module
One solution to solve this is to download the missing modules to a temporary path and then add that temporary path to PSModulePath when isa is run and remove the temp path when isa is finished running. But at this iteration this approach doesn't seem to work and needs further investigation.

The solution, that we pursue, is to download the missing module to a temp path. This temp path and its data is persistent in the sense that the temp path information is stored in $LOCALAPPDATA/PSScriptAnalyzer (pssaappdata). Whenver isa in invoked, it checks the temp path pointed in pssaappdata, checks if the modules are present in the temp directory and if so, copies them to the user psmodule path, which is typcially $HOME/Documents/WindowsPowerShell/Modules. And, when isa is finished running we remove those copied modues from the user psmodule path.