Interpret target in suppression as regular expression
According to README, target supports regular expression. By escaping the target argument, it is no longer a regular expression. This commit removes the escaping.

Ideally, we should not replace '*' with '.*' as we explicity mention that target takes regular expression. But, since globbing is more widely used in such scenario and if we do not interpret '*' as a glob, there is a chance it will invalidate suppression in many cases where people have used this feature. Hence, to maintain compatibility we keep '*' interpretation as a glob.