Add function to remove hyperlinks from markdown
The ReleaseMaker module extracts changelog for a given version from changelog.md and sets it 'ReleaseNotes' value in the module manifest. When we publish the module to powershellgallery.com, the release notes are displayed on the module page which cannot render the markdown links. Hence we need to remove them when plugging the changelog into release notes.