Add TextLines type for efficient line insertion and removal
Previous implementation used a List to maintain the lines of editable text. Inserting and removing at arbitrary points is expensive for Lists. So we use LinkedLists to store the lines, which makes inserting and removing lines a constant operation. However, this makes accessing a line through an index an expensive operation. The performance on this can easily be improved by maintaining a last accessed index/node.